Put flag group in own file.
