+++ src/components/Navigation/FlagGroup.scss
@@ -1,0 +1,16 @@
+.languageGroup {
+  display: flex;
+  justify-content: flex-end;
+  padding: 5px;
+
+  .languageItem {
+    align-self: center;
+    margin: 0 0 0 5px;
+    width: 30px;
+    height: 20px;
+    transition: all .075s ease-in-out;
+  }
+  .languageItem:hover {
+    transform: scale(1.2);
+  }
+}
