+++ src/components/Navigation/NavBar.scss
@@ -40,20 +40,3 @@ $color-search-border: #70aa66;
     }
   }
 }
-
-.languageGroup {
-  display: flex;
-  justify-content: flex-end;
-  padding: 5px;
-
-  .languageItem {
-    align-self: center;
-    margin: 0 0 0 5px;
-    width: 30px;
-    height: 20px;
-    transition: all .075s ease-in-out;
-  }
-  .languageItem:hover {
-    transform: scale(1.2);
-  }
-}
